+++ client/src/components/Loading/loading.module.css
@@ -1,0 +1,27 @@
+.container {
+  position: absolute;
+  background: #000;
+  color: #fff;
+  font-size: 35px;
+  width: 100%;
+  min-height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+}
+
+.spinner {
+  border: 7px solid #fff2;
+  width: 50px;
+  height: 50px;
+  border-radius: 50%;
+  border-left-color: #0af;
+
+  animation: spin 1s linear infinite;
+}
+
+@keyframes spin {
+  0% {transform: rotate(0deg)}
+  100% {transform: rotate(360deg)}
+}
