+++ client/src/components/PageAdmin/SearchBar/searchbar.module.css
@@ -1,6 +1,6 @@
 .box {
   display: flex;
-  justify-content: start;
+  justify-content: flex-start;
   margin-left: 1%;
 }
